[PATRO23-123]  Move classes to separate files
